+++ android/app/src/main/java/io/coursepick/coursepick/presentation/CoursePickApplication.kt
@@ -1,6 +1,8 @@
 package io.coursepick.coursepick.presentation
 
+import android.app.Activity
 import android.app.Application
+import android.os.Bundle
 import com.kakao.vectormap.KakaoMapSdk
 import io.coursepick.coursepick.BuildConfig
 import io.coursepick.coursepick.data.Services
@@ -19,13 +21,43 @@ class CoursePickApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
-        Logger.init(FirebaseAnalyticsService(installationId))
 
+        Logger.init(FirebaseAnalyticsService(installationId))
         if (BuildConfig.DEBUG) {
             Timber.plant(Timber.DebugTree())
         }
-
         KakaoMapSdk.init(this, BuildConfig.KAKAO_NATIVE_APP_KEY)
         CoursePickPreferences.init(this)
+        setUpCallbacks()
+    }
+
+    private fun setUpCallbacks() {
+        registerActivityLifecycleCallbacks(
+            object : ActivityLifecycleCallbacks {
+                override fun onActivityCreated(
+                    activity: Activity,
+                    savedInstanceState: Bundle?,
+                ) {
+                    Logger.log(Logger.Event.Enter(activity.javaClass.simpleName))
+                }
+
+                override fun onActivityStarted(activity: Activity) = Unit
+
+                override fun onActivityResumed(activity: Activity) = Unit
+
+                override fun onActivityPaused(activity: Activity) = Unit
+
+                override fun onActivityStopped(activity: Activity) = Unit
+
+                override fun onActivitySaveInstanceState(
+                    activity: Activity,
+                    outState: Bundle,
+                ) = Unit
+
+                override fun onActivityDestroyed(activity: Activity) {
+                    Logger.log(Logger.Event.Exit(activity.javaClass.simpleName))
+                }
+            },
+        )
     }
 }
