--- android/app/src/main/java/io/coursepick/coursepick/view/CoursePickApplication.kt
@@ -1,53 +1,0 @@
-package io.coursepick.coursepick.view
-
-import android.app.Activity
-import android.app.Application
-import android.os.Bundle
-import com.kakao.vectormap.KakaoMapSdk
-import io.coursepick.coursepick.BuildConfig
-import timber.log.Timber
-
-class CoursePickApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        if (BuildConfig.DEBUG) {
-            Timber.plant(Timber.DebugTree())
-        }
-        ClientIdProvider.init(this)
-        Logger.init(FirebaseLogHandler())
-        setUpCallbacks()
-
-        KakaoMapSdk.init(this, BuildConfig.KAKAO_NATIVE_APP_KEY)
-        CoursePickPreferences.init(this)
-    }
-
-    private fun setUpCallbacks() {
-        registerActivityLifecycleCallbacks(
-            object : ActivityLifecycleCallbacks {
-                override fun onActivityCreated(
-                    activity: Activity,
-                    savedInstanceState: Bundle?,
-                ) {
-                    Logger.log(Logger.Event.Enter(activity.javaClass.simpleName))
-                }
-
-                override fun onActivityStarted(activity: Activity) = Unit
-
-                override fun onActivityResumed(activity: Activity) = Unit
-
-                override fun onActivityPaused(activity: Activity) = Unit
-
-                override fun onActivityStopped(activity: Activity) = Unit
-
-                override fun onActivitySaveInstanceState(
-                    activity: Activity,
-                    outState: Bundle,
-                ) = Unit
-
-                override fun onActivityDestroyed(activity: Activity) {
-                    Logger.log(Logger.Event.Exit(activity.javaClass.simpleName))
-                }
-            },
-        )
-    }
-}
